FIX KB-26: Moved everything from `PokeBoxContent` into `PokeBoxPage`
