+++ frontend/toge-trades/src/Pages/Pokebox/PokeBoxPage.css
@@ -5,3 +5,32 @@
   margin-top:5em;
 
 }
+.box-content-container{
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+}
+.pokebox-search{
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+}
+.pokebox-cards{
+  margin-top: 2em;
+  display:grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap:1rem;
+  overflow-y:auto;
+  padding:1em;
+}
+.pokebox-loader{
+  justify-content: center;
+  align-items: center;
+}
+/* Mobile format */
+@media screen and (max-width: 768px) {
+  .pokebox-cards {
+    grid-template-columns: repeat(2, 1fr); 
+  }
+}
