--- frontend/toge-trades/src/components/pokebox/pokebox-contents/PokeBoxContents.css
@@ -1,26 +1,0 @@
-.box-content-container{
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-.pokebox-search{
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-}
-.pokebox-cards{
-  margin-top: 2em;
-  display:grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap:1rem;
-  overflow-y:auto;
-  padding:1em;
-}
-
-/* Mobile format */
-@media screen and (max-width: 768px) {
-  .pokebox-cards {
-    grid-template-columns: repeat(2, 1fr); 
-  }
-}
